+++ docs/changelog.md
@@ -8,8 +8,9 @@
 
 ## Версия 1.3.4 
 - Новые функции к Source: [getCategoryTracks](/func?id=getcategorytracks), [getListCategory](/func?id=getlistcategory).
-- Появился параметр [REQUESTS_IN_ROW](/guide?id=Параметры)
-- При чтении пустого файла через Cache.read выбрасывается исключение, чтобы предотвратить перезапись файла при баге со стороны Google ([подробнее](https://github.com/Chimildic/goofy/discussions/26))
+- Появился параметр [REQUESTS_IN_ROW](/guide?id=Параметры).
+- При чтении пустого файла через Cache.read выбрасывается исключение, чтобы предотвратить перезапись файла при баге со стороны Google ([подробнее](https://github.com/Chimildic/goofy/discussions/26)).
+- Новая функция [Playlist.saveWithUpdate](//func?id=savewithupdate).
 
 ## Версия 1.3.3
 - Оптимизация запросов к Last.fm в механизме накопления. Поиск только тех треков, что являются новыми для истории прослушиваний.